Client/requirements.txt
fastapi[all]
mysql-connector-python==8.0.30
python-dotenv==0.21.0
